+++ messages/de/admin.php
@@ -385,6 +385,8 @@ Da <strong>kein E-Mail-Versand</strong> eingerichtet ist, musst du die <strong>P
     'motion_type_initiator'       => 'Antragsteller*in / Unterstützer*innen: Anträge',
     'motion_type_initiator_amend' => 'Antragsteller*in / Unterstützer*innen: Änderungsanträge',
     'motion_type_proposed'        => 'Verfahrensvorschläge',
+    'motion_type_explan'          => 'Erklärung beim Anlegen',
+    'motion_type_explan_label'    => 'Eine abweichende Erklärung zum Anlegen von Anträgen angeben',
     'motion_type_proposed_label'  => 'Admins können Verfahrensvorschläge für Anträge festlegen',
     'motion_type_respons'         => 'Zuständigkeiten',
     'motion_type_respons_label'   => 'Jeder (Änderungs-)Antrag lässt sich einer Administrierenden zuordnen',