+++ models/settings/MotionType.php
@@ -7,9 +7,10 @@ class MotionType implements \JsonSerializable
     use JsonConfigTrait;
 
     /** @var string */
-    public $pdfIntroduction  = '';
-    public $cssIcon          = '';
-    public $motionTitleIntro = '';
+    public $pdfIntroduction   = '';
+    public $cssIcon           = '';
+    public $motionTitleIntro  = '';
+    public $createExplanation = '';
 
     /** @var bool */
     public $hasProposedProcedure = false;
@@ -17,4 +18,6 @@ class MotionType implements \JsonSerializable
     public $hasResponsibilities = false;
     /** @var bool */
     public $twoColMerging = false; // Can only be set manually as of yet
+    /** @var bool */
+    public $hasCreateExplanation = false;
 }