+++ views/admin/motion/type.php
@@ -22,6 +22,7 @@ $layout->addCSS('css/backend.css');
 $layout->loadSortable();
 $layout->loadDatepicker();
 $layout->loadFuelux();
+$layout->loadCKEditor();
 $layout->addAMDModule('backend/MotionTypeEdit');
 
 $myUrl = UrlHelper::createUrl(['admin/motion/type', 'motionTypeId' => $motionType->id]);
@@ -133,6 +134,28 @@ echo $controller->showErrors();
             $options = ['class' => 'form-control', 'id' => 'typeMotionIntro'];
             echo Html::textInput('type[typeMotionIntro]', $motionType->getSettingsObj()->motionTitleIntro, $options);
             ?>
+        </div>
+    </div>
+
+    <div class="form-group">
+        <div class="col-md-4 control-label">
+            <?= Yii::t('admin', 'motion_type_explan') ?>
+        </div>
+        <div class="col-md-8">
+            <?php
+            echo HTMLTools::labeledCheckbox(
+                'type[hasCreateExplanation]',
+                Yii::t('admin', 'motion_type_explan_label'),
+                $motionType->getSettingsObj()->hasCreateExplanation,
+                'typeHasCreateExplanation'
+            );
+            ?>
+            <div id="typeCreateExplanationHolder" class="wysiwyg-textarea" data-full-html="0">
+                <textarea title="<?= Html::encode(Yii::t('admin', 'motion_type_explan')) ?>" name="type[createExplanation]"
+                    ><?= Html::encode($motionType->getSettingsObj()->createExplanation) ?></textarea>
+                <div class="texteditor motionTextFormattings boxed" id="typeCreateExplanation_wysiwyg"
+                    ><?= $motionType->getSettingsObj()->createExplanation ?></div>
+            </div>
         </div>
     </div>