+++ views/motion/edit_form.php
@@ -41,7 +41,9 @@ echo '<div class="form content hideIfEmpty">';
 echo $controller->showErrors();
 
 $publicPolicies = [IPolicy::POLICY_ALL, IPolicy::POLICY_LOGGED_IN, IPolicy::POLICY_WURZELWERK];
-if (in_array($form->motionType->policyAmendments, $publicPolicies)) {
+if ($form->motionType->getSettingsObj()->hasCreateExplanation) {
+    echo $form->motionType->getSettingsObj()->createExplanation . '<br><br>';
+} elseif (in_array($form->motionType->policyAmendments, $publicPolicies)) {
     echo '<div style="font-weight: bold; text-decoration: underline;">' .
         Yii::t('motion', 'create_explanation_title') . '</div>' .
         str_replace('%HOME%', UrlHelper::homeUrl(), Yii::t('motion', 'create_explanation')) .