+++ History.md
@@ -5,6 +5,7 @@
 - The pink deadline circle on the consultation is now also shown if multiple motion types with the same deadline exist.
 - First, small beginnings of a REST API are implemented, currently with read-only access. It can be enabled in the site component settings. The documentation can be found at [docs/openapi.yaml](docs/openapi.yaml)
 - Tabular data sections now skip empty fields when displaying them.
+- In the "Create a motion" form, every motion type can now have an own explanation text at the top of the page, instead of the default "Motion or amendment?" one.
 - Internal: the login system now supports plugins for retrieving user accounts from external sources, e.g. CMS systems with an existing user database. As an example, a integration into Drupal/CiviCRM can be found in the plugins/drupal_civicrm-folder.
 - Internal: Plugins can now provide custom amendment numberings and add extra settings and data fields for amendments and motions.